Refactor all maze generating algorithms to borrow Grid instead of owning.
Besides the fact that Aldous-Broder and Wilson's don't even make use of
mutably owning the Grid struct, all algorithms only link the `GridCell`s
which are Rc<RefCell<T>>, thus Grid itself is not mutated.

Additionally, this removes the need for cloning Grid or additional cloning
of `GridCell`s (as was the case in SideWinder and BinaryTree) which means
a much smaller memory footprint.